Fix value converter

+++ src/ImageHotspot/Models/ImageHotspotValue.cs
@@ -4,8 +4,8 @@ namespace ImageHotspot.Models;
 
 public class ImageHotspotValue
 {
-    public int Left { get; set; }
-    public int Top { get; set; }
+    public decimal Left { get; set; }
+    public decimal Top { get; set; }
     public decimal PercentX { get; set; }
     public decimal PercentY { get; set; }
     public int Width { get; set; }
@@ -14,6 +14,10 @@ public class ImageHotspotValue
 
     public override string ToString()
     {
-        return "left: " + PercentX.ToString(CultureInfo.InvariantCulture) + "%; top: " + PercentY.ToString(CultureInfo.InvariantCulture) + "%;";
+        return "left: "
+            + PercentX.ToString(CultureInfo.InvariantCulture)
+            + "%; top: "
+            + PercentY.ToString(CultureInfo.InvariantCulture)
+            + "%;";
     }
 }
